+++ posix_spawn.c
@@ -112,11 +112,11 @@ int posix_spawn(pid_t *pid, const char *path,
 
 	/* Do not proceed to any hack if not in chroot */
 	if (!__inchroot()) {
-		__warn_if_not_preloading_libiamroot();
+		__note_if_not_preloading_libiamroot_and_ensure_preloading();
 		__execfd();
-		__verbose_exec(path, argv, envp);
+		__verbose_exec(path, argv, __environ);
 		return next_posix_spawn(pid, path, file_actions, attrp, argv,
-					envp);
+					__environ);
 	}
 
 	ret = __interpreter_script(program, argv, buf, sizeof(buf), off,